chore: migrate chat.getMentionedMessages endpoint to new OpenAPI pattern with AJV validation
