+++ .changeset/migrate-chat-getMentionedMessages-to-openapi.md
@@ -1,0 +1,6 @@
+---
+'@rocket.chat/meteor': minor
+'@rocket.chat/rest-typings': minor
+---
+
+Migrated chat.getMentionedMessages endpoint to new OpenAPI pattern with AJV validation
