+++ packages/rest-typings/src/v1/chat.ts
@@ -562,39 +562,6 @@ const GetPinnedMessagesSchema = {
 
 export const isChatGetPinnedMessagesProps = ajv.compile<GetPinnedMessages>(GetPinnedMessagesSchema);
 
-type GetMentionedMessages = {
-	roomId: IRoom['_id'];
-	count?: number;
-	offset?: number;
-	sort?: string;
-};
-
-const GetMentionedMessagesSchema = {
-	type: 'object',
-	properties: {
-		roomId: {
-			type: 'string',
-			minLength: 1,
-		},
-		count: {
-			type: 'number',
-			nullable: true,
-		},
-		offset: {
-			type: 'number',
-			nullable: true,
-		},
-		sort: {
-			type: 'string',
-			nullable: true,
-		},
-	},
-	required: ['roomId'],
-	additionalProperties: false,
-};
-
-export const isChatGetMentionedMessagesProps = ajv.compile<GetMentionedMessages>(GetMentionedMessagesSchema);
-
 type ChatSyncMessages = {
 	roomId: IRoom['_id'];
 	lastUpdate?: string;
@@ -956,14 +923,6 @@ export type ChatEndpoints = {
 	};
 	'/v1/chat.getPinnedMessages': {
 		GET: (params: GetPinnedMessages) => {
-			messages: IMessage[];
-			count: number;
-			offset: number;
-			total: number;
-		};
-	};
-	'/v1/chat.getMentionedMessages': {
-		GET: (params: GetMentionedMessages) => {
 			messages: IMessage[];
 			count: number;
 			offset: number;
